F/paypal subscription

+++ apps/web-shop-front2/src/graphql/order/useCreateWageIntent.ts
@@ -1,6 +1,6 @@
 import { gql, useMutation } from '@apollo/client';
 const mutation = gql`
-  mutation CreateWageIntent($amount: numeric!, $currency: String = "USD", $order_id: uuid!) {
+  mutation CreateWageIntent($amount: Float!, $currency: String = "USD", $order_id: uuid!) {
     createWageIntent(data: { amount: $amount, currency: $currency, order_id: $order_id }) {
       link
     }
